ocrservice_rule-dev/requirements.txt
opencv-python
pytesseract
scipy
werkzeug==0.14.1
pillow==5.1.0
attrs==19.1.0
flask==0.12.2
numpy
pymongo
